+++ src/model/BaseOrganizationItem.js
@@ -362,7 +362,7 @@ BaseOrganizationItem.prototype['update_time'] = undefined;
 BaseOrganizationItem.prototype['visible_to'] = undefined;
 
 /**
- * The ID of the label
+ * The label assigned to the organization
  * @member {Number} label
  */
 BaseOrganizationItem.prototype['label'] = undefined;
@@ -374,7 +374,7 @@ BaseOrganizationItem.prototype['label'] = undefined;
 BaseOrganizationItem.prototype['owner_name'] = undefined;
 
 /**
- * The BCC email of the organization
+ * The BCC email associated with the organization
  * @member {String} cc_email
  */
 BaseOrganizationItem.prototype['cc_email'] = undefined;
@@ -627,7 +627,7 @@ BaseOrganizationItemFields.prototype['update_time'] = undefined;
  */
 BaseOrganizationItemFields.prototype['visible_to'] = undefined;
 /**
- * The ID of the label
+ * The label assigned to the organization
  * @member {Number} label
  */
 BaseOrganizationItemFields.prototype['label'] = undefined;
@@ -637,7 +637,7 @@ BaseOrganizationItemFields.prototype['label'] = undefined;
  */
 BaseOrganizationItemFields.prototype['owner_name'] = undefined;
 /**
- * The BCC email of the organization
+ * The BCC email associated with the organization
  * @member {String} cc_email
  */
 BaseOrganizationItemFields.prototype['cc_email'] = undefined;