+++ src/model/OrganizationItem.js
@@ -377,7 +377,7 @@ OrganizationItem.prototype['update_time'] = undefined;
 OrganizationItem.prototype['visible_to'] = undefined;
 
 /**
- * The ID of the label
+ * The label assigned to the organization
  * @member {Number} label
  */
 OrganizationItem.prototype['label'] = undefined;
@@ -389,7 +389,7 @@ OrganizationItem.prototype['label'] = undefined;
 OrganizationItem.prototype['owner_name'] = undefined;
 
 /**
- * The BCC email of the organization
+ * The BCC email associated with the organization
  * @member {String} cc_email
  */
 OrganizationItem.prototype['cc_email'] = undefined;
@@ -660,7 +660,7 @@ BaseOrganizationItemWithEditNameFlag.prototype['update_time'] = undefined;
  */
 BaseOrganizationItemWithEditNameFlag.prototype['visible_to'] = undefined;
 /**
- * The ID of the label
+ * The label assigned to the organization
  * @member {Number} label
  */
 BaseOrganizationItemWithEditNameFlag.prototype['label'] = undefined;
@@ -670,7 +670,7 @@ BaseOrganizationItemWithEditNameFlag.prototype['label'] = undefined;
  */
 BaseOrganizationItemWithEditNameFlag.prototype['owner_name'] = undefined;
 /**
- * The BCC email of the organization
+ * The BCC email associated with the organization
  * @member {String} cc_email
  */
 BaseOrganizationItemWithEditNameFlag.prototype['cc_email'] = undefined;