+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The file format of it is based on [Keep a Changelog](http://keepachangelog.com/e
 For public Changelog covering all changes done to Pipedrive’s API, webhooks and app extensions platforms, see [public Changelog](https://pipedrive.readme.io/docs/changelog) with discussion area in [Developers Community](https://devcommunity.pipedrive.com/c/documentation/changelog/19).
 
 ## [Unreleased]
+### Added
+- Added `GET /organizations/collection` endpoint to fetch all organizations with cursor pagination
+- Added `GET /persons/collection` endpoint to fetch all persons with cursor pagination
 
 ## [21.0.0] - 2023-07-27
 ### Fixed